+++ packages/react-relay/jest.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+	testEnvironment: 'jsdom',
+	testRegex: '/tests/.*\\.test\\.tsx?$',
+	setupFilesAfterEnv: ['<rootDir>/jest-setup.ts'],
+	transform: {
+		'^.+\\.(t|j)sx?$': '@swc/jest',
+	},
+	coveragePathIgnorePatterns: ['/node_modules/', '/dist/'],
+	coverageReporters: ['text', 'html'],
+	reporters: ['default', 'github-actions'],
+};