+++ packages/nextjs/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.10](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.10.9...@wundergraph/nextjs@0.10.10) (2023-04-19)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
+## [0.10.9](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.10.8...@wundergraph/nextjs@0.10.9) (2023-04-19)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
+## [0.10.8](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.10.7...@wundergraph/nextjs@0.10.8) (2023-04-19)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
+## [0.10.7](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.10.6...@wundergraph/nextjs@0.10.7) (2023-04-17)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
+## [0.10.6](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.10.5...@wundergraph/nextjs@0.10.6) (2023-04-15)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
+## [0.10.5](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.10.4...@wundergraph/nextjs@0.10.5) (2023-04-14)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
+## [0.10.4](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.10.3...@wundergraph/nextjs@0.10.4) (2023-04-14)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
 ## [0.10.3](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.10.2...@wundergraph/nextjs@0.10.3) (2023-04-13)
 
 **Note:** Version bump only for package @wundergraph/nextjs